pyproject.toml
[tool.poetry]
name = "tp-django"
version = "0.1.0"
description = ""
authors = ["Antonio Palma Pernia <palmapernia@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.8"
mysqlclient = "^2.2.7"
pillow = "^12.0.0"
django-debug-toolbar = "^6.1.0"
djangorestframework = "^3.16.1"
djangorestframework-simplejwt = "^5.5.1"
drf-spectacular = "^0.29.0"
gunicorn = "^23.0.0"
whitenoise = "^6.11.0"


[tool.poetry.group.dev.dependencies]
coverage = "^7.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = [
    "*/__pycache__/",
    "*/__init__.py",
    "*/wsgi.py",
    "*/asgi.py",
    "*/migrations/",
    "manage.py",
    "app/urls.py"
]
